backend/requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
redis
pydantic
httpx
python-dotenv
groq 
PyPDF2>=3.0.0
python-docx>=1.2.0
langchain-community>=0.0.21